Patch to Imakefile.
(1) Add sed command to take out "Pixmap" from beginning of lines in
app-defaults file.

(2) Change /net/koala/... to $(INCDIR), etc.  This was causing a very
strange effect on my machine because (a) my amd is sitting on /net with
a "hosts" map, and (b) there is a machine called "koala" on the net! ;)