Bring in the modifications to SUP that I did for Berkeley's TCS.
1) Preserve the timestamp of an updated file even if a command that is
executed because of the update modifies the file.  This allows you to
do things like export /etc/hosts and when it changes, customize it based
on the target host with a simple script, but not upgrade it every time
sup is run.

2) Add the "unlinkbusy" SUP option.  The goal of this option is to allow
you to upgrade binaries and libraries on a running system via sup
automagically just like "install" does.

3) Add the "rename" collection function.  This allows you to store a file
on the server under one name, but have it land on the client under another.

All of these changes are targeted toward making SUP a better administrative
tool.  A reverse rdist if you like.  TCS uses sup to keep over 100 HP
workstations in sync with a master machine and performs all maintainence
including new software installations and system patches using it without
a hitch.  Of course, the SUP collections are fairly complicated
(HUP inetd when inetd.conf changes is a very simple example), but it
definitely works and works well.

This patch also includes full support for the HPUX platform.  I didn't
see any reason to strip it out since HPUX was the target platform of
the original work and it may be of use to other people.