Fix "version required" comment.
How do you think I found this out?  It's a long story, but here's the
short version (no pun intended):

I had this upgraded to 0.5 on my machine, and had sent it to Chuck for
a review.  Then Peter did an upgrade.  So I did a "cvs update", which
showed one conflict (date in patch, understandable).

However, when I did a "cvs diff", it showed that Peter had forgotten
something.  So here it is. :)