Remove NO_PACKAGE, since this is what the source says:
===
JOE is free software; you can redistribute it and/or modify it under the
terms of the GNU General Public License as published by the Free Software
Foundation; either version 1, or (at your option) any later version.
===                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Since it's not (only) under GPL1, there shouldn't be any problem
with us distributing it in any medium.