This is the networking-enhanced version of DGD.
The following additional functionality is provided:

send and receive udp datagrams
initiate tcp and telnet connections
listen on any number of ports for incoming connections

This is a separate version because sites not requiring these features can
use the other version and not need to concern themselves with the network
security issues. Also the programming semantics for network communication
are somewhat different in the two versions.