This is the Perl5 equivalent of the MNT_NOSUID patch that was put into /usr/bin/perl a while ago. It works with 5.003 and 5.004-tobe.
I can't test it on freefall (I don't have a setuid script to play with)
but it works at home. I'll test on thud later.

Every setuid script found on a partition mounted "nosuid" will not be executed
and "Permission Denied" will be displayed.

It won't be in soo-to-be-released 5.004 because it is not a FreeBSD speciific
problem but it will be dealt with in 5.004_01 (maintenance release).

Obtained from: Jörg's patch (edited for Perl5 by me).