Upgrade to version 0.89.
NOTE, this is now the full version.  Michael has pushed the PGP hooks
development off-shore.  This port now fetches from there.  Mutt does
not contain crypo, just hooks to run an extern application.  All I've
talked to feel ITAR does NOT prohibit this -- as Mutt isn't using the
PGP ABI.  In fact, our current Elm package is built with PGP hooks and
we do not control it, so I am treaing Mutt in the same way.