Upgrade to the first pre-release of KDE 1.1.
This port requireat least s version 1.41 of the Qt library.
There will be an error reported by configure, if only an earlier
version is found, but no automatic port dependency exists (i.e.
the x11-toolkits/qt141 port has to be manually built and installed).

There may still be a problem with a missing -lXext in the kdesupport
port. This will be taken care of during the next few days, if the
problem still exists ...