Update to lesstif 0.88.9.
Note that this change to the Makefile is a format change and the
real commit was rev 1.33. All of the others are the actual update.
The message got lost in an accidental ctrl-c during the commit.
Thanks to jdp for fixing all the locks.

PR:		12630
Submitted by:	Steven G. Kargl <kargl@troutmask.apl.washington.edu>