Make my marker match reality.
If you'll believe this, my original entry was a guess.
I looked at an online map that had vancouver, but not Abbotsford
(where I live) so I moved the mouse over a bit and took that value.
It was:
49.01  -122.68
I finally got around to getting the correct value, and it now is:
49.05  -122.48