From maintainer:
The way the hevea port is installed now, if it is
installed as a package, then chances are hevea.sty
will not be visible to latex.  What I have done is -
instead of having the port copy hevea.sty to a place
where latex can find it, I give instructions as a
message for the user to put it there.

PR:		16149
Submitted by:	maintainer