Persuade libxml2 to install into different filesystem space than libxml (cursing the author all the way for having two incompatible versions of their software do this).
This is still not 100%, but a step in the right direction for those
ports that need libxml2, but also have to co-exist with GNOME
(which, for the moment, uses libxml exclusively).

Max will be committing a port shortly to show exactly how to
tell a port to use libxml2 under this new scheme of things.

As and when GNOME (the primary consumer of libxml) moves over
to libxml2, these hacks will be reverted, and similar hacks put
in to libxml until it dies a quiet death.

Submitted by:	sobomax