Add dlmalloc 2.7.0, a very efficient malloc/free/realloc implementation.
According to my tests it is *amazingly* efficient - it gave me about 10%
memory saving (SIZE in the top(1)) for the large processes like X and jre
without any measureable performance saturation.

Moreover, due to not very clear for me reasons Python benchmark (pybench)
is about 60% (!!!) faster with this allocator comparing to the libc one.
Obviously we should investigate this further and if there is no error
then tune either Python or our own malloc.

Alternatively we may evaluate a possibility to make it (or part of it) our
default malloc(), because it is friendly licensed and actively maintained.