Expand on the existing documentation regarding TLS and certificates, adding information important to operators of SMTP services used by a number of Windows (and probably Unix) MUA packages.
This text has been approved by the author and will be included in
the next release of exim-4.  Another release of exim-3.3x is not
expected.

This change is based on an explanation of SSL certificates attributed
below, but was not a direct submission.  Errors are my own, etc.

Submitted by:	terry
Message-Id:	<3C3F3A93.C1ECF9B0@mindspring.com>