This port now installs below ${PREFIX}/share/doc, where it should.
The tutorial is not specific to JDK 1.1.8, it pertains to
Java versions up to 1.4. That's why I've changed the PORTVERSION
from 1.1.8 to 2002.03.04, the date the tutorial was last updated.
See:
http://java.sun.com/docs/books/tutorial/information/download.html

Since this is not really a JDK tutorial but rather a Java
tutorial, the port installation directory is now called
java-tutorial.

Not displaying the message anymore. It is evident where the files
go.

Changed the pkg-plist to use %%T%% instead of
jdk%%VERSION%%/docs/books/tutorial as the prefix.