Roll mpeg-lib into the two gimp ports.  mpeg-lib is now abandoned ware, and will be removed from the ports tree soon.
mpeg-lib will now be built solely for the purpose of gimp, and no libmpeg.*
will be installed on the system.  This will avoid a namespace conflict with
KDE which also installs a libmpeg.so.  These two libraries are incompatible.

Submitted by:	alane and myself
Not reviewed by:	alane
Inspired by:	mail/evolution