Major overhaul to reflect the update to the latest version, 1.2.1.
Development stalled on this distribution for a while, then went into a
frantic period of many updates per month. Now it's stabilized to the
point that it's worth keeping current again.

There are many new models of UPS supported, and a lot of changes to
how the various tools work, so users of the old port should read UPGRADING.

Unfortunately, the CGI stuff depends on a version of gd that we don't have
yet, so that'll have to be sidelined for a while.