Update exiscan patch to -24, to catch up with the protocol change in p5-Mail-SpamAssassin-2.50.
That port update may have been ill-advised so close to a ports freeze,
but this new version of the exiscan patch is backward compatible to
the original protocol, so if the maintainer of p5-Mail-SpamAssassin
rolls back, this port will not require a change.

Bump PORTREVISION accordingly.

Requested by:	"David Haworth" <dave@fyonn.net>