Make libstdc++-2.96-112.7.1.i386.rpm to be extracted and included into package.  libstdc++-2.96 is a core package that is referenced by many linux_base applications such as linux-eagle.  For instance, linux-eagle doesn't work with linux_base-7.1_3.
Despite the importance, libstdc++-2.96 has been forgotten to be extracted
since 6 months.  (Actually, the RPM package is listed in the Makefile,
but was badly defined so that no extraction was performed.)

Since linux_base is an important package, I hesitated to commit this
correction (for a long time :).  If there is any incovenience in the
way of package definition, please feel free to correct my change.